# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_ICreateAdmin.username-or-name_": {
				"properties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ResAdmin": {
				"$ref": "#/components/schemas/Pick_ICreateAdmin.username-or-name_"
			},
			"Pick_ICreateAdmin.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"AdminResDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ICreateAdmin.name_"
					},
					{
						"properties": {
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"_id"
						],
						"type": "object"
					}
				]
			},
			"ICreateAdmin": {
				"properties": {
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ICreateAdmin.username-or-password_": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"AdminLogin": {
				"$ref": "#/components/schemas/Pick_ICreateAdmin.username-or-password_"
			},
			"UserAuthenticate": {
				"properties": {
					"token": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILoginUser": {
				"properties": {
					"lPhone": {
						"type": "string"
					},
					"lPassword": {
						"type": "string"
					}
				},
				"required": [
					"lPhone",
					"lPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ICreateUser.email_": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IEmailVerify": {
				"$ref": "#/components/schemas/Pick_ICreateUser.email_"
			},
			"Pick_ICreateUser.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserResDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ICreateUser.name_"
					},
					{
						"properties": {
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"_id"
						],
						"type": "object"
					}
				]
			},
			"ICreateUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"phone",
					"password",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignJWT": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"phone",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "BE",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin": {
			"get": {
				"operationId": "GetAdminById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/ResAdmin"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/AdminResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateAdmin"
							}
						}
					}
				}
			}
		},
		"/admin/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminLogin"
							}
						}
					}
				}
			}
		},
		"/authenticate/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserAuthenticate"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILoginUser"
							}
						}
					}
				}
			}
		},
		"/authenticate/send-code-email": {
			"post": {
				"operationId": "SendVerifyCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmailVerify"
							}
						}
					}
				}
			}
		},
		"/authenticate/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateUser"
							}
						}
					}
				}
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateUser"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/IUserInfo"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/sign-jwt-token": {
			"post": {
				"operationId": "SignJwtToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignJWT"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "FindUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}